#3224 - Added support for running modules as part of expressions that reference them - Introduced ExpressionRunnerService which has access to expressions, modules, and can download source code as needed - Standardised how modules and the built in Map module are pulled in (Map is done on startup though) - Removed dependency from quantifications on running a built-in expression to get the columns for spectrum eV calibration - Version was null in code-editor HTML, added a check for it. This is not necessarily the right solution (?) - Code editor wasn't copying moduleReferences to the new expression it created - Implemented crude caching in data module service, it was querying the API for every module get - Added a check to expression editor to not try to bring up help for Lua language files because it was showing lots of unnecessary errors
Need to fully test data module and expression caching behaviour of the source code field, ensure we don't re-request it once we have it (maybe implement something time-based though?). Will also need to do further testing and optimisation of how we run our expressions.

Need to figure out a neater way to init Lua WASM - was trying to do it from service constructor but this ended up with race conditions or double-initing.